Challenges. Finally.
This took way too fucking long.
Anyway, despite there almost definitely being some bugs left in this, as well as some balancing that needs to be done, I'm releasing the update so that you guys can get started on challenge 9. Because it still sucks.
Also added a "Max all" button because everyone was complaining.